+++ instrumentation/opentelemetry-instrumentation-redis/src/opentelemetry/instrumentation/redis/util.py
@@ -26,11 +26,11 @@ def _extract_conn_attributes(conn_kwargs):
     attributes["db.name"] = db
     attributes["db.redis.database_index"] = db
     try:
-        attributes["net.peer.name"] = conn_kwargs["host"]
-        attributes["net.peer.ip"] = conn_kwargs["port"]
+        attributes["net.peer.name"] = conn_kwargs.get("host", "localhost")
+        attributes["net.peer.port"] = conn_kwargs.get("port", 6379)
         attributes["net.transport"] = "IP.TCP"
     except KeyError:
-        attributes["net.peer.name"] = conn_kwargs["path"]
+        attributes["net.peer.name"] = conn_kwargs.get("path", "")
         attributes["net.transport"] = "Unix"
 
     return attributes