+++ CHANGELOG.md
@@ -20,6 +20,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   ([#424](https://github.com/open-telemetry/opentelemetry-python-contrib/pull/424))
 - Update gRPC instrumentation to better wrap server context
   ([#420](https://github.com/open-telemetry/opentelemetry-python-contrib/pull/420))
+- `opentelemetry-instrumentation-redis` Fix default port KeyError and Wrong Attribute name (net.peer.ip -> net.peer.port)
+  ([#265](https://github.com/open-telemetry/opentelemetry-python-contrib/pull/265))
+- `opentelemetry-instrumentation-asyncpg` Fix default port KeyError and Wrong Attribute name (net.peer.ip -> net.peer.port)
+  ([#265](https://github.com/open-telemetry/opentelemetry-python-contrib/pull/265))
 
 ### Added
 - `opentelemetry-instrumentation-urllib3` Add urllib3 instrumentation